Add support for referencing other playlists from auto playlists
This adds two new conditions for auto playlists:

    * "is in playlist"
    * "is not in playlist"

These allow an auto playlist to filter songs based on other playlists,
both manual and auto.

Note that there is currently no checking for circular dependencies (e.g.
playlist A depending on B depending on A), but in practice it will hit
the maximum recursion depth very quickly when trying to render out the
library.